python/momentum_persistence/requirements.txt
pandas>=2.0
numpy>=1.24
yfinance>=0.2.40
supabase>=2.0
python-dotenv>=1.0
